Remove types.d file
This particular implementation is an external interface.  As such it
should not be in a .d.ts file but instead part of the interface.
